3-langchain/hello_langchain.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv
import os

load_dotenv()
google_api_key = os.getenv("GEMINI_API_KEY")

def hello_langchain():
    llm = ChatGoogleGenerativeAI(model="gemini-2.5-flash", google_api_key=google_api_key)
    response = llm.invoke("Roast python developers")
    print(response.content)

if __name__ == "__main__":
    hello_langchain()